// src/Types/Types.ts
export type bookType = {
    picture: string
    name: string
    year: string
    genre: string
    author: string
    id: number
}


export type sendedBookType = {
    picture: string
    name: string
    year: string
    genre: string
    author: string
}

export type loginType = {
   login: string
   password: string
}